sidecar/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
requests==2.32.3
chromadb==0.5.5
pydantic==2.9.2
pyyaml>=6.0
qrcode[pil]>=7.4
openai==1.82.0
